refactor: Rename UseSocket to SingleViewSocket
