+++ frontend/src/components/View/Single/SingleViewSocket.js
